Change App Name And Switch To Viewsets
