+++ cityservices/urls.py
@@ -18,5 +18,5 @@ from django.conf.urls import url, include
 
 
 urlpatterns = [
-    url(r'^', include('streetsweeper.urls')),
+    url(r'^', include('streets.urls')),
 ]
