--- streetsweeper/views.py
@@ -1,49 +1,0 @@
-from django.contrib.auth.models import User
-from rest_framework import (
-    generics,
-    permissions,
-)
-from rest_framework.decorators import api_view
-from rest_framework.response import Response
-from rest_framework.reverse import reverse
-
-from streetsweeper.models import Street
-from streetsweeper.permissions import IsOwnerOrReadOnly
-from streetsweeper.serializers import (
-    StreetSerializer,
-    UserSerializer,
-)
-
-
-@api_view(['GET'])
-def api_root(request, format=None):
-    return Response({
-        'users': reverse('user-list', request=request, format=format),
-        'streets': reverse('street-list', request=request, format=format)
-    })
-
-
-class StreetList(generics.ListCreateAPIView):
-    queryset = Street.objects.all()
-    serializer_class = StreetSerializer
-    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
-
-    def perform_create(self, serializer):
-        serializer.save(owner=self.request.user)
-
-class StreetDetail(generics.RetrieveUpdateDestroyAPIView):
-    queryset = Street.objects.all()
-    serializer_class = StreetSerializer
-    permission_classes = (
-        permissions.IsAuthenticatedOrReadOnly,
-        IsOwnerOrReadOnly)
-
-
-class UserList(generics.ListAPIView):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-
-
-class UserDetail(generics.RetrieveAPIView):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
